+++ lib/flutter_donation_buttons.dart
@@ -1,4 +1,6 @@
 library flutter_donation_buttons;
+
 export './donationButtons/paypalButton.dart';
 export './donationButtons/ko-fiButton.dart';
 export './donationButtons/patreonButton.dart';
+export './donationButtons/buyMeACoffeButton.dart';