+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 ## 0.0.1
 Added Kofi, PayPal and Patreon Button
+
+## 0.0.2
+Fixed some small staff in the documentation, tried to format the data better
+
+## 0.0.3
+- Added example folder and main.dart
+- Added [Buy me a coffe](https://www.buymeacoffee.com) button
+- Added images